Prototype

+++ run.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 
-# Usage: run.sh <compacting|copying|no|generational|incremental> <performance|limit> <scenario_name>
+# Usage: run.sh <compacting|copying|no|generational|incremental> <performance|limit|live> <scenario_name>
 
 if [ -z "$1" ] || [ -z "$2" ] || [ -z "$3" ]
 then
@@ -12,9 +12,9 @@ then
     echo "First argument must be 'compacting', 'copying', 'no', 'generational', or 'incremental'"
     exit 1
 fi
-if [ "$2" != "performance" ] && [ "$2" != "limit" ]
+if [ "$2" != "performance" ] && [ "$2" != "limit" ] && [ "$2" != "live" ]
 then
-    echo "Second argument must be 'performance' or 'limit'"
+    echo "Second argument must be 'performance', 'limit', or 'live'"
     exit 1
 fi
 GC_FLAG=--$1-gc
@@ -22,19 +22,33 @@ if [ "$2" == "performance" ]
 then
     DFXTEMPLATE=performance-dfx.json
     CANISTER=benchmark
-else 
+fi
+if [ "$2" == "limit" ]
+then 
     DFXTEMPLATE=limit-dfx.json
     CANISTER=limit-tester
+fi
+if [ "$2" == "live" ]
+then 
+    DFXTEMPLATE=live-dfx.json
+    CANISTER=benchmark
 fi
 awk '// {gsub("#SCENARIO#", "'$3'"); gsub("#GCFLAG#", "'$GC_FLAG'"); print }' $DFXTEMPLATE > dfx.json
 dfx start --clean --background
 dfx deploy
-dfx canister call $CANISTER run "()"
-if [ $? != 0 ]
+if [ "$2" == "live" ] 
 then
-    echo "Canister call failed"
+    dfx canister call $CANISTER run "()"&
+    npm start
+else
+    dfx canister call $CANISTER run "()"
+    if [ $? != 0 ]
+    then
+        echo "Canister call failed"
+        dfx stop
+        exit 1
+    fi
     dfx stop
-    exit 1
+    rm dfx.json
 fi
-dfx stop
-rm dfx.json
+
